refactor: relocated logout server action
