+++ src/app/(authenticated)/logout/actions.ts
